Add files via upload
C14_v1, stands for version one, contains the work being done on a logistic regression model trained to predict sentiment and a BART summarization tool applied on large data set. Where as in version 2, the same regression model is loaded and tested on custom created dataset and also applied the same summarization tool with tuning.